Added RoomController, Service, and Exception handlers

+++ HostelManagementTeam05/src/main/java/com/avsoft/hostelmanagement/repostiory/FloorRepository.java
@@ -9,7 +9,7 @@ import com.avsoft.hostelmanagement.entity.Floor;
 
 
 
-public interface FloorRepository extends JpaRepository<Floor, Integer>  {
+public interface FloorRepository extends JpaRepository<Floor, Long>  {
 
 	List<GetFloorDto> findByBuildingId(int id);
 }
